# docs/openapi.yaml
openapi: 3.0.1
info:
  title: EBSI-like Trusted Registry API
  description: "This OpenAPI specification defines the REST API endpoints for interacting\
    \ with the DOME EBSI-like Trusted Registry, including adapted JSON-RPC methods\
    \ for blockchain transactions as REST API endpoints. "
  version: 4.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /v4/schemas/{schemaId}:
    get:
      tags:
      - Trusted Schemas Registry
      summary: Get details of a specific registered Schema
      operationId: getSchema
      parameters:
      - name: SchemaId
        in: path
        description: The ID of the Schema to retrieve.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Details of the Schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaDetails'
        "404":
          description: No Schema has the requested ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WrongRequest'
    put:
      tags:
      - Trusted Schemas Registry
      summary: Update an already registered Schema
      operationId: updateSchema
      parameters:
      - name: SchemaId
        in: path
        description: The ID of the Schema to update
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaDetails'
        required: true
      responses:
        "200":
          description: Schema updated successfully.
        "404":
          description: No Schema has the requested ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WrongRequest'
    delete:
      tags:
      - Trusted Schemas Registry
      summary: Deletes a specific registered Schema
      operationId: deleteSchema
      parameters:
      - name: schemaId
        in: path
        description: The ID of the schema to delete.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The Schema was deleted succesfully
  /v4/participants/{participantId}:
    get:
      tags:
      - Trusted Participants Registry
      summary: Get details of a specific participant
      operationId: getParticipant
      parameters:
      - name: participantId
        in: path
        description: The DID of the participant to retrieve.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Participant details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantDetails'
        "404":
          description: No participant has the requested ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WrongRequest'
    put:
      tags:
      - Trusted Participants Registry
      summary: Update an existing participant
      operationId: updateParticipant
      parameters:
      - name: participantId
        in: path
        description: The DID of the participant to update.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipantDetails'
        required: true
      responses:
        "200":
          description: Transaction built successfully.
        "400":
          description: No participant has the requested ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WrongRequest'
    delete:
      tags:
      - Trusted Participants Registry
      summary: Deletes a specific registered Participant
      operationId: deleteParticipant
      parameters:
      - name: participantId
        in: path
        description: The DID of the participant to delete.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The Participant was deleted succesfully
...
  /v4/schemas:
    get:
      tags:
      - Trusted Schemas Registry
      summary: List all registered Schemas
      operationId: listSchemas
      parameters:
      - name: "page[after]"
        in: query
        description: Cursor for pagination (starting point)
        required: false
        schema:
          type: integer
          format: int32
      - name: "page[size]"
        in: query
        description: Number of items per page
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: A list of registered Schemas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSchemas200Response'
    post:
      tags:
      - Trusted Schemas Registry
      summary: Register a new Schema associated with a public key
      operationId: registerSchema
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaDetails'
        required: true
      responses:
        "200":
          description: Schema registered successfully.
  /v4/participants:
    get:
      tags:
      - Trusted Participants Registry
      summary: List all trusted participants
      operationId: listParticipants
      parameters:
      - name: "page[after]"
        in: query
        description: Cursor for pagination (starting point)
        required: false
        schema:
          type: integer
          format: int32
      - name: "page[size]"
        in: query
        description: Number of items per page
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: A list of trusted participants with pagination details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListParticipants200Response'
    post:
      tags:
      - Trusted Participants Registry
      summary: Insert a new participant
      operationId: insertParticipant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipantDetails'
        required: true
      responses:
        "200":
          description: Participant was registered successfully.
...
components:
  schemas:
    WrongRequest:
      required:
      - errCode
      - errMessage
      type: object
      properties:
        errCode:
          type: integer
          description: The HTTP status code
          format: int32
        errMessage:
          type: string
          description: A message describing the error
...
    SchemaDetails:
      required:
      - id
      - schemaData
      type: object
      properties:
        id:
          type: string
          description: The identifier of this item.
        schemaData:
          type: string
          description: The JSON representation of the Schema
    Attribute:
      required:
      - body
      - hash
      - issuerType
      type: object
      properties:
        hash:
          type: string
        issuerType:
          type: string
        body:
          $ref: '#/components/schemas/AttributeBody'
      description: LEAR Credential Issuer's attributes
    AttributeBody:
      required:
      - claims
      - credentialsType
      - validFor
      type: object
      properties:
        credentialsType:
          type: string
        validFor:
          $ref: '#/components/schemas/AttributeValidity'
        claims:
          type: array
          items:
            $ref: '#/components/schemas/AttributeClaim'
    AttributeClaim:
      required:
      - allowedValues
      - name
      type: object
      properties:
        name:
          type: string
        allowedValues:
          type: array
          items:
            $ref: '#/components/schemas/AttributeClaimAllowedValue'
    AttributeClaimAllowedValue:
      required:
      - description
      - value
      type: object
      properties:
        value:
          type: string
        description:
          type: string
    AttributeValidity:
      required:
      - from
      - to
      type: object
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
    ParticipantDetails:
      required:
      - did
      type: object
      properties:
        did:
          type: string
          description: The Decentralized Identifier (DID) of this participant.
        attributes:
          type: array
          description: Empty list
          items:
            $ref: '#/components/schemas/Attribute'
          default: []
        hasAttributes:
          type: boolean
          description: Whether the attributes list is empty or not
          default: false
...
    List200ResponseLinks:
      required:
      - first
      - last
      - next
      - prev
      type: object
      properties:
        first:
          type: string
        prev:
          type: string
        next:
          type: string
        last:
          type: string
...
    TrustedRegistrySummary:
      required:
      - did
      - href
      type: object
      properties:
        did:
          type: string
          description: The Decentralized Identifier (DID) of this item.
        href:
          type: string
          description: The link to the details of this item.
    ListSchemas200Response:
      required:
      - items
      - links
      - pageSize
      - self
      - total
      type: object
      properties:
        self:
          type: string
          description: The URL of the current page of results.
        items:
          type: array
          items:
            $ref: '#/components/schemas/TrustedRegistrySummary'
        total:
          type: integer
          description: The total number of access nodes in the registry.
          format: int32
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
        links:
          $ref: '#/components/schemas/List200ResponseLinks'
    ListParticipants200Response:
      required:
      - items
      - links
      - pageSize
      - self
      - total
      type: object
      properties:
        self:
          type: string
          description: The URL of the current page of results.
        items:
          type: array
          items:
            $ref: '#/components/schemas/TrustedRegistrySummary'
        total:
          type: integer
          description: The total number of access nodes in the registry.
          format: int32
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
        links:
          $ref: '#/components/schemas/List200ResponseLinks'
...